Move random number generator to util file
